Crate domain_block_preprocessor

This crate provides a preprocessor for the domain block, which is used to construct domain extrinsics from the consensus block.

The workflow is as follows:

Extract domain-specific bundles from the consensus block.

Compile the domain bundles into a list of extrinsics.

Shuffle the extrisnics using the seed from the consensus chain.

Filter out the invalid xdm extrinsics.

Push back the potential new domain runtime extrisnic.

Modules

inherents

Provides functionality of adding inherent extrinsics to the Domain. Unlike Primary chain where inherent data is first derived the block author and the data is verified by the on primary runtime, domains inherents short circuit the derivation and verification of inherent data as the inherent data is directly taken from the primary block from which domain block is being built.

stateless_runtime

Structs

DomainBlockPreprocessor

PreprocessResult

Traits

ValidateReceipt