domain_block_preprocessor

Module inherents

Provides functionality of adding inherent extrinsics to the Domain.

Unlike Primary chain where inherent data is first derived the block author and the data is verified by the on primary runtime, domains inherents short circuit the derivation and verification of inherent data as the inherent data is directly taken from the primary block from which domain block is being built.

One of the first use case for this is passing Timestamp data. Before building a domain block using a primary block, we take the current time from the primary runtime and then create an unsigned extrinsic that is put on top the bundle extrinsics.

Deriving these extrinsics during fraud proof verification should be possible since verification environment will have access to consensus chain.

Structs

CreateInherentDataProvider

Functions

extract_domain_runtime_upgrade_code

Returns new upgraded runtime if upgraded did happen in the provided consensus block.

get_inherent_data