Struct domain_block_preprocessor::runtime_api_light::RuntimeApiLight

pub struct RuntimeApiLight<Executor> { /* private fields */ }

This is a stateless wrapper around the runtime api. Note: Dispatch may give invalid output if the invoked function relies on state of the chain. Note: Be sure to use this only when there is no domain client available as this holds since entire domain runtime on heap.

Implementations

impl<Executor> RuntimeApiLight<Executor>where Executor: CodeExecutor,

pub fn new(executor: Arc<Executor>, runtime_code: Cow<'static, [u8]>) -> Self