domain_block_preprocessor::stateless_runtime

Struct StatelessRuntime

pub struct StatelessRuntime<CBlock, Block, Executor> { /* private fields */ }

Stateless runtime api based on the runtime code and partial state if provided.

NOTE:

This is only supposed to be used when no domain client available, i.e., when the caller does not own the entire domain state.

This perfectly fits the runtime APIs that are purely stateless, but itโs also usable for the stateful APIs. If some states are used inside a runtime api, these states must be provided and set before dispatching otherwise StatelessRuntime may give invalid output.

Implementations

impl<CBlock, Block, Executor> StatelessRuntime<CBlock, Block, Executor>
where CBlock: BlockT, Block: BlockT, Executor: CodeExecutor + RuntimeVersionOf,

pub fn new(executor: Arc<Executor>, runtime_code: Cow<'static, [u8]>) -> Self

Create a new instance of StatelessRuntime with empty storage.

pub fn set_storage(&mut self, storage: Storage)

Set the storage.

Inject the state necessary for calling stateful runtime APIs.

pub fn set_extension_factory( &mut self, extension_factory: Box<dyn ExtensionsFactory<Block>>, )

Set the extensions.

Inject the necessary extensions for Domain.

pub fn outbox_storage_key( &self, message_key: MessageKey, ) -> Result<Vec<u8>, ApiError>

pub fn inbox_response_storage_key( &self, message_key: MessageKey, ) -> Result<Vec<u8>, ApiError>

pub fn channel_storage_key( &self, chain_id: ChainId, channel_id: ChannelId, ) -> Result<Vec<u8>, ApiError>

pub fn extract_signer( &self, extrinsics: Vec<Block::Extrinsic>, ) -> Result<ExtractSignerResult<Block>, ApiError>

pub fn construct_set_code_extrinsic( &self, runtime_code: Vec<u8>, ) -> Result<Vec<u8>, ApiError>

pub fn construct_timestamp_extrinsic( &self, moment: Moment, ) -> Result<Block::Extrinsic, ApiError>

pub fn construct_consensus_chain_byte_fee_extrinsic( &self, consensus_chain_byte_fee: Balance, ) -> Result<Block::Extrinsic, ApiError>

pub fn construct_domain_update_chain_allowlist_extrinsic( &self, updates: DomainAllowlistUpdates, ) -> Result<Block::Extrinsic, ApiError>

pub fn is_inherent_extrinsic( &self, extrinsic: &Block::Extrinsic, ) -> Result<bool, ApiError>

pub fn find_first_inherent_extrinsic( &self, extrinsics: &Vec<Block::Extrinsic>, ) -> Result<Option<u32>, ApiError>

pub fn is_xdm_mmr_proof_valid( &self, extrinsic: &Block::Extrinsic, ) -> Result<Option<bool>, ApiError>

pub fn extract_xdm_mmr_proof( &self, extrinsic: &Block::Extrinsic, ) -> Result<Option<Vec<u8>>, ApiError>

pub fn batch_extract_native_xdm_mmr_proof( &self, extrinsics: &Vec<Block::Extrinsic>, ) -> Result<BTreeMap<u32, ConsensusChainMmrLeafProof<NumberFor<CBlock>, CBlock::Hash, H256>>, ApiError>

pub fn extract_native_xdm_mmr_proof( &self, extrinsic: &Block::Extrinsic, ) -> Result<Option<ConsensusChainMmrLeafProof<NumberFor<CBlock>, CBlock::Hash, H256>>, ApiError>

pub fn decode_extrinsic( &self, opaque_extrinsic: OpaqueExtrinsic, ) -> Result<Result<Block::Extrinsic, DecodeExtrinsicError>, ApiError>

pub fn decode_extrinsics_prefix( &self, opaque_extrinsics: Vec<OpaqueExtrinsic>, ) -> Result<Vec<Block::Extrinsic>, ApiError>

pub fn is_within_tx_range( &self, extrinsic: &Block::Extrinsic, bundle_vrf_hash: &U256, tx_range: &U256, ) -> Result<bool, ApiError>

pub fn extract_signer_if_all_within_tx_range( &self, extrinsics: &Vec<Block::Extrinsic>, bundle_vrf_hash: &U256, tx_range: &U256, ) -> Result<Result<Vec<Option<AccountId>>, u32>, ApiError>

pub fn check_extrinsics_and_do_pre_dispatch( &self, uxts: Vec<Block::Extrinsic>, block_number: NumberFor<Block>, block_hash: Block::Hash, ) -> Result<Result<(), CheckExtrinsicsValidityError>, ApiError>

This is a stateful runtime api call and requires setting storage keys.

pub fn transfers_storage_key(&self) -> Result<Vec<u8>, ApiError>

pub fn block_fees_storage_key(&self) -> Result<Vec<u8>, ApiError>

pub fn extrinsic_weight( &self, extrinsic: &Block::Extrinsic, ) -> Result<Weight, ApiError>

pub fn extrinsics_weight( &self, extrinsics: &Vec<Block::Extrinsic>, ) -> Result<Weight, ApiError>

pub fn is_valid_sudo_call(&self, extrinsic: Vec<u8>) -> Result<bool, ApiError>

pub fn construct_domain_sudo_extrinsic( &self, inner_call: Vec<u8>, ) -> Result<Block::Extrinsic, ApiError>

pub fn construct_evm_contract_creation_allowed_by_extrinsic( &self, inner_call: PermissionedActionAllowedBy<EthereumAccountId>, ) -> Result<Block::Extrinsic, ApiError>