Type Alias domain_block_preprocessor::stateless_runtime::ExtractedStateRoots

pub type ExtractedStateRoots<Block> = ExtractedStateRootsFromProof<NumberFor<Block>, <Block as BlockT>::Hash, <Block as BlockT>::Hash>;

Aliased Type

struct ExtractedStateRoots<Block> {
    pub consensus_chain_block_info: BlockInfo<<<Block as Block>::Header as Header>::Number, <Block as Block>::Hash>,
    pub consensus_chain_state_root: <Block as Block>::Hash,
    pub domain_info: Option<(DomainId, BlockInfo<<<Block as Block>::Header as Header>::Number, <Block as Block>::Hash>, <Block as Block>::Hash)>,
}

Fields
consensus_chain_block_info: BlockInfo<<<Block as Block>::Header as Header>::Number, <Block as Block>::Hash>

Consensus chain block info when proof was constructed

consensus_chain_state_root: <Block as Block>::Hash

State root of Consensus chain at above number and block hash.

domain_info: Option<(DomainId, BlockInfo<<<Block as Block>::Header as Header>::Number, <Block as Block>::Hash>, <Block as Block>::Hash)>

Storage proof that src chain state_root is registered on Consensus chain. This is optional when the src_chain is the consensus chain. BlockNumber and BlockHash is used with storage proof to validate and fetch its state root.