Struct domain_block_preprocessor::runtime_api_light::RuntimeApiLight

pub struct RuntimeApiLight<Executor> { /* private fields */ }

Lightweight runtime api based on the runtime code and partial state.

NOTE:

This is only supposed to be used when no domain client available, i.e., when the caller does not own the entire domain state.

This perfectly fits the runtime APIs that are purely stateless, but it’s also usable for the stateful APIs. If some states are used inside a runtime api, these states must be provided and set before dispatching otherwise RuntimeApiLight may give invalid output.

Implementations

impl<Executor> RuntimeApiLight<Executor>where Executor: CodeExecutor,

pub fn new(executor: Arc<Executor>, runtime_code: Cow<'static, [u8]>) -> Self

Create a new instance of RuntimeApiLight with empty storage.

pub fn set_storage(&mut self, storage: Storage)

Set the storage.

Inject the state necessary for calling stateful runtime APIs.